Use statistics from Suggest command for badge numbers
This patch updates the badge generation process to use the objects determined
by Inch::API::Suggest (instead of all objects in the codebase).

The reasons:

* the badges should mimick the output of `inch suggest` rather than `inch stats`
* to encourage people to document their code, low priority objects should
  not count against the overall assessment of the badge

After this change, low priority objects (like private methods and objects marked
as :nodoc: or @private) are no longer relevant for the badge generator.
(They are however still displayed in the status page, which continues to give
the complete picture.)